+++ .vscode/launch.json
@@ -1,0 +1,14 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Connect to okteto",
+            "type": "go",
+            "request": "attach",
+            "mode": "remote",
+            "remotePath": "/usr/src/app",
+            "port": 2345,
+            "host": "127.0.0.1"
+        }
+    ]
+}
