+++ bashrc
@@ -1,0 +1,5 @@
+cat << EOF
+Welcome to your development container. Happy coding!
+EOF
+
+export PS1="\[\e[36m\]\${OKTETO_NAMESPACE:-okteto}:\[\e[32m\]\${OKTETO_NAME:-dev} \[\e[m\]\W> "
