+++ chart/templates/NOTES.txt
@@ -1,0 +1,1 @@
+Success! Your application will be available shortly.
