+++ chart/values.yaml
@@ -1,0 +1,2 @@
+replicaCount: 1
+image: ""
