+++ pkg/database/database.go
@@ -1,0 +1,36 @@
+package database
+
+import (
+	"database/sql"
+	"fmt"
+
+	_ "github.com/lib/pq"
+)
+
+const (
+	host     = "postgresql"
+	port     = 5432
+	user     = "okteto"
+	password = "okteto"
+	dbname   = "votes"
+)
+
+func Open() *sql.DB {
+	psqlconn := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable", host, port, user, password, dbname)
+	for {
+		db, err := sql.Open("postgres", psqlconn)
+		if err == nil {
+			return db
+		}
+	}
+}
+
+func Ping(db *sql.DB) {
+	fmt.Println("Waiting for postgresql...")
+	for {
+		if err := db.Ping(); err == nil {
+			fmt.Println("Postgresql connected!")
+			return
+		}
+	}
+}
