--- rent/.idea/libraries/Maven__ch_qos_logback_logback_classic_1_2_6.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: ch.qos.logback:logback-classic:1.2.6">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/ch/qos/logback/logback-classic/1.2.6/logback-classic-1.2.6.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/ch/qos/logback/logback-classic/1.2.6/logback-classic-1.2.6-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/ch/qos/logback/logback-classic/1.2.6/logback-classic-1.2.6-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
