--- rent/.idea/libraries/Maven__com_fasterxml_jackson_core_jackson_core_2_12_5.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.fasterxml.jackson.core:jackson-core:2.12.5">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-core/2.12.5/jackson-core-2.12.5.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-core/2.12.5/jackson-core-2.12.5-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-core/2.12.5/jackson-core-2.12.5-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
