+++ rent/.stignore
@@ -1,0 +1,3 @@
+.git
+?
+*.jar
