--- rentals/chart/Chart.yaml
@@ -1,7 +1,0 @@
-apiVersion: v2
-name: movies-rentals
-description: Rentals backend for Movies App
-type: application
-version: 0.1.0
-appVersion: 1.0.0
-icon: https://apps.okteto.com/movies/icon.png
