--- rentals/chart/templates/NOTES.txt
@@ -1,1 +1,0 @@
-Success! Your application will be available shortly.
