--- rentals/chart/templates/api-deployment.yaml
@@ -1,27 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: api
-  labels:
-    {{- include "movies.labels" . | nindent 4 }}
-    app.kubernetes.io/component: api
-spec:
-  replicas: {{ .Values.replicaCount }}
-  selector:
-    matchLabels:
-      app.kubernetes.io/name: api
-      app.kubernetes.io/instance: {{ .Release.Name }}
-  template:
-    metadata:
-      labels:
-        app.kubernetes.io/name: api
-        app.kubernetes.io/instance: {{ .Release.Name }}
-    spec:
-      terminationGracePeriodSeconds: 0
-      containers:
-        - name: api
-          image: {{ .Values.image }}
-          ports:
-            - name: http
-              containerPort: 8080
-              protocol: TCP
