--- rentals/chart/templates/api-ingress.yaml
@@ -1,19 +1,0 @@
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: api
-  labels:
-    {{- include "movies.labels" . | nindent 4 }}
-  annotations:
-    dev.okteto.com/generate-host: "movies"
-spec:
-  rules:
-    - http:
-        paths:
-          - path: /rentals
-            pathType: Prefix
-            backend:
-              service:
-                name: api
-                port:
-                  number: 8080
