--- rentals/chart/templates/api-service.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: api
-  labels:
-    {{- include "movies.labels" . | nindent 4 }}
-spec:
-  type: ClusterIP
-  ports:
-    - port: 8080
-      targetPort: 8080
-      protocol: TCP
-      name: http
-  selector:
-    app.kubernetes.io/name: api
-    app.kubernetes.io/instance: {{ .Release.Name }}
