--- rentals/chart/templates/worker-deployment.yaml
@@ -1,22 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    app: worker
-  name: worker
-spec:
-  replicas: {{ .Values.replicaCount }}
-  selector:
-    matchLabels:
-      app: worker
-  template:
-    metadata:
-      labels:
-        app: worker
-    spec:
-      containers:
-      - image: {{ .Values.image }}
-        name: worker
-        command:
-          - /usr/local/bin/worker
-
