--- rentals/chart/values.yaml
@@ -1,2 +1,0 @@
-replicaCount: 1
-image: ""
