--- rentals/pkg/kafka/kafka.go
@@ -1,28 +1,0 @@
-package kafka
-
-import (
-	"fmt"
-
-	kingpin "gopkg.in/alecthomas/kingpin.v2"
-
-	"github.com/Shopify/sarama"
-)
-
-var (
-	brokerList = kingpin.Flag("brokerList", "List of brokers to connect").Default("kafka:9092").Strings()
-)
-
-func GetMaster() sarama.Consumer {
-	kingpin.Parse()
-	config := sarama.NewConfig()
-	config.Consumer.Return.Errors = true
-	brokers := *brokerList
-	fmt.Println("Waiting for kafka...")
-	for {
-		master, err := sarama.NewConsumer(brokers, config)
-		if err == nil {
-			fmt.Println("Kafka connected!")
-			return master
-		}
-	}
-}
